fix: move the duplicate session check to the replay graph session datagen
